Draft for demonstrating multi-release in `ConcurrencyUtil`
Related-to: 836ddc49071070877aad3ea8d520c6391f771b52
